do some fix and finsh the category crud

+++ src/Controller/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Model\CategoryManager;
 use App\Model\ArticleManager;
+use App\Service\CategoryService;
 
 class CategoryController extends AbstractController
 {
@@ -38,16 +39,19 @@ class CategoryController extends AbstractController
 
     public function addCategory(): string
     {
-        // $errors = [];
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $category = array_map('trim', $_POST);
+            $categoryService = new CategoryService();
+            $categoryService->categoryTitleErrors($category);
+            $categoryService->categoryPhotoErrors($category);
 
-            // TODO security
-
-            $categoryManager = new CategoryManager();
-            $categoryManager->insertCategory($category);
-            header('Location:/');
-            die();
+            if (empty($categoryService->errors)) {
+                $categoryManager = new CategoryManager();
+                $categoryManager->insertCategory($category);
+                header('Location:/');
+                die();
+            }
+            return $this->twig->render('Article/addCategory.html.twig', ['errors' => $categoryService->errors]);
         }
         if (isset($_SESSION['admin']) === true) {
             return $this->twig->render('Article/addCategory.html.twig');
@@ -62,20 +66,22 @@ class CategoryController extends AbstractController
         $categoryManager = new categoryManager();
         $category = $categoryManager->selectOneById($id);
 
-        // $errors = [];
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $category = array_map('trim', $_POST);
+            $categoryService = new CategoryService();
+            $categoryService->categoryTitleErrors($category);
+            $categoryService->categoryPhotoErrors($category);
 
-            // TODO security
-
-            $categoryManager->updateCategory($category);
-            header('Location:/category/list');
-            die();
-
-            // return $this->twig->render('Category/editCategory.html.twig', [
-            //     'category' => $category,
-            //     'errors' => $categoryService->errors
-            // ]);
+            if (empty($categoryService->errors)) {
+                $categoryManager = new CategoryManager();
+                $categoryManager->updateCategory($category);
+                header('Location:/category/list');
+                die();
+            }
+            return $this->twig->render('Article/editCategory.html.twig', [
+                'category' => $category,
+                'errors' => $categoryService->errors
+            ]);
         }
 
         if (isset($_SESSION['admin']) === true) {
@@ -88,14 +94,26 @@ class CategoryController extends AbstractController
         }
     }
 
-    public function delete(): void
+    public function delete()
     {
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $id = trim($_POST['id']);
             $categoryManager = new CategoryManager();
-            $categoryManager->delete((int)$id);
-
-            header('Location:/category/list');
+            $articleManager = new ArticleManager();
+            if (empty($articleManager->selectArticlesByCategory((int)$id))) {
+                $categoryManager->delete((int)$id);
+                header('Location:/category/list');
+                die();
+            } else {
+                $categoryManager = new CategoryManager();
+                $category = $categoryManager->selectAll();
+                $error = "Vous avez toujours des articles dans cette catégorie,
+                Veuillez supprimer ou changer de catégorie les articles";
+                return $this->twig->render('Article/categoryList.html.twig', [
+                    'category' => $category,
+                    'error' => $error
+                ]);
+            }
         }
     }
 }
